+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Added
 ~~~~~
 
 - Added :meth:`Client.fetch_creative_island()`.
+- Added ``{current_playlist}`` variable to client's status.
 
 
 v0.0.3